+++ src/test/resources/inspections/ArrayMapToCollectionInspection/namespaced_array_map.php
@@ -3,6 +3,6 @@
 namespace MyAwesomeName\Good;
 
 $a = [1, 2, 3, 4, 5];
-$b = collect($a)->map(function ($n) {
+$b = <caret>\array_map(function($n) {
     return ($n * $n * $n);
-})->all();
+}, $a);